backend/requirements.txt


fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings


sqlalchemy>=2.0.25
alembic
asyncpg
python-dotenv
redis
psycopg2-binary

langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai
langgraph
docker
requests
langchain-ollama

python-multipart
pypdf
pillow